requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
autopep8==2.3.1
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coverage==7.6.0
crispy-bootstrap5==0.7
cryptography==42.0.3
cssselect2==0.7.0
defusedxml==0.8.0rc2
diff-match-patch==20230430
distlib==0.3.8
dj-rest-auth==5.0.2
Django==4.2.15
django-allauth==0.60.1
django-bootstrap-v5==1.0.11
django-browser-reload==1.11.0
django-chartjs==2.3.0
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-debug-toolbar==4.4.6
django-formtools==2.5.1
django-formwizard==1.0
django-htmx==1.17.2
django-import-export==3.3.7
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
filelock==3.13.1
flake8==7.1.0
html5lib==1.1
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
lxml==5.3.0
MarkupPy==1.14
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.2
oscrypto==1.3.0
packaging==24.1
pandas==2.2.1
pathspec==0.12.1
phonenumbers==8.13.32
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.5.0
psycopg2==2.9.9
pycodestyle==2.12.0
pycparser==2.21
pyflakes==3.2.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==5.0.1
pytest==8.3.2
pytest-django==4.8.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==8.0
reportlab==4.2.5
requests==2.31.0
requests-oauthlib==1.3.1
ruff==0.4.4
setuptools==69.0.3
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.3.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.1
virtualenv==20.25.0
webencodings==0.5.1
wfastcgi==3.0.0
whitenoise==6.7.0
xhtml2pdf==0.2.16
xlrd==2.0.1
xlwt==1.3.0
